redo refactroing socket io
